refactor(core): unify to single mdast-based pipeline (#97)
Remove the multipass streaming pipeline in favor of the MDAST-based
approach. This eliminates the `MARKFLOW_PIPELINE=mdast` environment
variable toggle and consolidates all rendering through the AST-first
architecture.

Key changes:

- Add `enable_lazy_images` and `enable_smartypants` options to MDAST
  renderer, matching features previously only in multipass pipeline
- Extract `jsx_normalize` module from markdown_adapter for reuse
- Remove multipass pipeline files: adapter.rs, engine.rs, event.rs,
  span.rs, html.rs, streaming_rewriter.rs, markdown_adapter.rs
- Update NAPI bindings: remove `parse`, `parse_with_options` functions,
  add new options to `BlockOptions`
- Update WASM bindings: remove `render_html`, `stream_html` functions
- Clean up tests that depended on multipass API

The MDAST pipeline now handles:
- Directive preprocessing (:::note, etc.)
- JSX component rendering
- Lazy image loading (loading="lazy" attribute)
- Smart punctuation (smartypants)
- Heading extraction for TOC

Closes #97
