+++ crates/core/src/parser/mod.rs
@@ -1,5 +1,4 @@
 #![allow(missing_docs)]
-pub mod markdown_adapter;
 pub mod parse_config;
 
 pub use parse_config::{ParseConfig, ParseConstructs};
