+++ crates/core/src/renderer/mod.rs
@@ -1,9 +1,6 @@
 #![allow(missing_docs)]
-pub mod html;
 pub mod mdast;
-pub mod streaming_rewriter;
 
 pub use mdast::{
     BlocksResult, HeadingEntry, Options as MdastOptions, PropValue, RenderBlock, to_blocks,
 };
-pub use streaming_rewriter::{RewriteOptions, StreamingRewriter};
