+++ crates/core/src/transform/mod.rs
@@ -1,21 +1,17 @@
-//! Streaming transform stage utilities (event adapters).
+//! Transform utilities for markdown processing.
 //!
-//! - `hoist_adapter`: lifts top-level ESM import/export statements from the event stream.
-//! - `directive_adapter`: rewrites `:::` directives into configured block tags during streaming.
 //! - `code_fence`: tracks fenced blocks to avoid hoisting/rewrites inside them.
+//! - `directives`: directive mapping traits and default implementations.
+//! - `jsx_normalize`: JSX indentation normalization for MDAST rendering.
+//! - `smartypants`: smart punctuation transformations (quotes, dashes, ellipsis).
 
 /// Code fence state tracking utilities.
 pub mod code_fence;
 /// Astro docs component rewrite helpers.
 pub mod components;
-/// Directive-to-Aside (or custom) rewriting adapter.
-pub mod directive_adapter;
 /// Directive mapping traits and default implementations.
 pub mod directives;
-/// ESM import/export hoisting adapter.
-pub mod hoist_adapter;
+/// JSX indentation normalization for MDAST rendering.
+pub mod jsx_normalize;
 /// Smart punctuation transformations (quotes, dashes, ellipsis).
 pub mod smartypants;
-
-pub use directive_adapter::DirectiveAdapter;
-pub use hoist_adapter::HoistAdapter;
