--- crates/core/tests/integration_tests.rs
@@ -1,141 +1,0 @@
-use markflow_core::{RewriteOptions, parse, parse_with_options};
-use std::fs;
-use std::path::PathBuf;
-
-// This helper was in the original lib.rs test module.
-fn fixtures_dir() -> PathBuf {
-    PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../fixtures/core")
-}
-
-// This helper was also in the original lib.rs test module.
-fn read_fixture(path: &str) -> String {
-    fs::read_to_string(fixtures_dir().join(path)).unwrap()
-}
-
-#[test]
-fn test_parse_applies_lazy_loading() {
-    let input = "![alt](img.png)";
-    let output = parse(input).unwrap().html;
-    assert!(output.contains("loading=\"lazy\""));
-}
-
-#[test]
-fn test_parse_without_hoist_keeps_imports_in_html() {
-    let input = "import X from './x'\n\n# Title";
-    let opts = RewriteOptions {
-        enable_hoist: false,
-        enable_smartypants: false,
-        ..RewriteOptions::default()
-    };
-    let result = parse_with_options(input, opts).expect("parse succeeds");
-
-    assert!(
-        result.html.contains("import X from './x'"),
-        "import should remain in HTML when hoist is disabled"
-    );
-    assert!(
-        result.imports.is_empty(),
-        "hoisted import list should be empty when hoist is disabled"
-    );
-}
-
-#[test]
-fn test_parse_table_alignment_and_math() {
-    let input = "| A | B |\n|:-|:-:|\n| $x$ | $$y$$ |";
-    let output = parse(input).unwrap().html;
-    assert!(output.contains("<table>"));
-    assert!(
-        output.contains("<td style=\"text-align:left\"><span class=\"math-inline\">x</span></td>")
-    );
-    assert!(output.contains("<span class=\"math-inline\">y</span>"));
-}
-
-#[test]
-fn test_parse_frontmatter_passthrough() {
-    let input = "---\ntitle: test\n---\n\ncontent";
-    let output = parse(input).unwrap().html;
-    assert!(output.contains("frontmatter"));
-    assert!(output.contains("title: test"));
-}
-
-#[test]
-fn test_reference_link_resolves_definition() {
-    let input = "[Example][ref]\n\n[ref]: https://example.com \"Example Site\"";
-    let output = parse(input).unwrap().html;
-    assert!(output.contains("<a href=\"https://example.com\""));
-    assert!(output.contains("title=\"Example Site\""));
-}
-
-#[test]
-fn test_smartypants_disabled() {
-    let input = "Hello -- \"world\" ...";
-    let opts = RewriteOptions {
-        enable_smartypants: false,
-        ..RewriteOptions::default()
-    };
-    let output = parse_with_options(input, opts).unwrap().html;
-    assert!(
-        output.contains("Hello -- &quot;world&quot; ...")
-            || output.contains("Hello -- \"world\" ...")
-    );
-}
-
-#[test]
-fn test_directive_rewrites_to_aside() {
-    let input = ":::note[Heads up]\ncontent\n:::";
-    let output = parse(input).unwrap().html;
-    assert!(output.contains("<aside class=\"aside aside--note\">"));
-    assert!(output.contains("Heads up"));
-    assert!(output.contains("</aside>"));
-}
-
-#[test]
-fn test_reference_image_resolves_definition() {
-    let input = "![Alt][logo]\n\n[logo]: https://cdn.example.com/logo.png \"Logo\"";
-    let output = parse(input).unwrap().html;
-    assert!(output.contains("<img src=\"https://cdn.example.com/logo.png\""));
-    assert!(output.contains("title=\"Logo\""));
-}
-
-#[test]
-fn test_mdx_embedded_jsx_preserved() {
-    let input = read_fixture("mdx/embedded-jsx/component.mdx");
-    let output = parse(&input).unwrap().html;
-    assert!(output.contains("<aside class=\"aside\">"));
-    assert!(output.contains("Heads up"));
-    assert!(output.contains("</aside>"));
-}
-
-#[test]
-fn test_mdx_inline_expression_preserved() {
-    let input = read_fixture("mdx/expressions/inline.mdx");
-    let output = parse(&input).unwrap().html;
-    assert!(
-        output.contains("{props.name ?? 'friend'}"),
-        "output: {output}"
-    );
-}
-
-#[test]
-fn test_mdx_flow_expression_preserved() {
-    let input = read_fixture("mdx/expressions/flow.mdx");
-    let output = parse(&input).unwrap().html;
-    assert!(output.contains("steps.join(' → ');"), "output: {output}");
-}
-
-#[test]
-fn test_mdx_esm_import_preserved() {
-    let input = read_fixture("mdx/esm/imports.mdx");
-    let output = parse(&input).unwrap();
-    assert!(
-        !output.html.contains("import Tabs from"),
-        "root-level imports should be hoisted away from HTML output"
-    );
-    assert!(
-        output
-            .imports
-            .iter()
-            .any(|spec| spec.source.contains("import Tabs from")),
-        "imports should be hoisted into the imports list"
-    );
-}
